refactor: Extract core combinators
