+++ src/lib.rs
@@ -434,6 +434,7 @@ pub mod lib {
 }
 
 pub use self::bits::*;
+pub use self::core::*;
 pub use self::internal::*;
 pub use self::traits::*;
 
@@ -443,6 +444,7 @@ pub use self::str::*;
 pub mod error;
 
 pub mod combinator;
+mod core;
 mod internal;
 mod traits;
 #[macro_use]
